Change phase of channel 1 as well
For some reason when phase of channel 2 was changed in positive direction,
this would cause a glitch when polarity was inverted, this doesn't
happen in the negative direction.

To avoid said glitch, if we require a positive step, we instead change
channel 1 to move in the negative direction, this in effect is like
having a possitive step in channel 2.

Additionally, a software prescaler was added to slow down the movement.